fix
60647 [Timer-1] ERROR uk.co.la1tv.websiteUploadProcessor.JobPoller  -
SQLExcepti           on when trying to query databases for files that need
processing.
com.mysql.jdbc.exceptions.jdbc4.MySQLDataException: '2.237565592E9' in
column '6           ' is outside valid range for the datatype INTEGER.
        at sun.reflect.NativeConstructorAccessorImpl.newInstance0(Native
	Method)
	        at
		sun.reflect.NativeConstructorAccessorImpl.newInstance(NativeConstruct
		orAccessorImpl.java:57)
		        at
			sun.reflect.DelegatingConstructorAccessorImpl.newInstance(DelegatingC
			onstructorAccessorImpl.java:45)
			        at
				java.lang.reflect.Constructor.newInstance(Constructor.java:526)
				        at
					com.mysql.jdbc.Util.handleNewInstance(Util.java:406)
					        at
						com.mysql.jdbc.Util.getInstance(Util.java:381)
						        at
							com.mysql.jdbc.SQLError.createSQLException(SQLError.java:999)
							        at
								com.mysql.jdbc.SQLError.createSQLException(SQLError.java:956)
								        at
									com.mysql.jdbc.SQLError.createSQLException(SQLError.java:926)
									        at
										com.mysql.jdbc.ResultSetImpl.throwRangeException(ResultSetImpl.java:7
										819)
										        at
											com.mysql.jdbc.ResultSetImpl.parseIntAsDouble(ResultSetImpl.java:7062
											)
											        at
												com.mysql.jdbc.ResultSetImpl.getInt(ResultSetImpl.java:2623)
												        at
													com.mysql.jdbc.ResultSetImpl.getInt(ResultSetImpl.java:2734)
													        at
														uk.co.la1tv.websiteUploadProcessor.helpers.DbHelper.buildFileFromResu
														lt(DbHelper.java:44)
														        at
															uk.co.la1tv.websiteUploadProcessor.JobPoller$PollTask.processFiles(Jo
															bPoller.java:87)
															        at
																uk.co.la1tv.websiteUploadProcessor.JobPoller$PollTask.run(JobPoller.j
																ava:65)
																        at
																	java.util.TimerThread.mainLoop(Timer.java:555)
																	        at
																		java.util.TimerThread.run(Timer.java:505)

+++ src/main/java/uk/co/la1tv/websiteUploadProcessor/File.java
@@ -21,10 +21,10 @@ public class File {
 	
 	private int id;
 	private String name;
-	private int size;
+	private double size;
 	private FileTypeAbstract type;
 	
-	public File(int id, String name, int size, FileTypeAbstract type) {
+	public File(int id, String name, double size, FileTypeAbstract type) {
 	
 		this.id = id;
 		this.name = name;
@@ -51,7 +51,7 @@ public class File {
 		return name;
 	}
 	
-	public int getSize() {
+	public double getSize() {
 		return size;
 	}
